+++ packages/field-help-icon/jest.config.js
@@ -1,7 +1,0 @@
-const global = require('../../jest.config.global');
-
-module.exports = {
-  ...global,
-  displayName: 'field-help-icon',
-  coverageDirectory: '../../coverage/field-help-icon',
-};