+++ packages/paper/CHANGELOG.md
@@ -8,8 +8,7 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ## 0.1.0 (2023-05-22)
 
-
 ### Features
 
-* **mui-paper:** add paper component ([4bf7776](https://github.com/Availity/element/commit/4bf7776a0f19e3566983e176901233a6292fa745))
-* **mui-paper:** refactor how props are spread ([ecddf69](https://github.com/Availity/element/commit/ecddf6908fdb69418705cb854aa84c419de3ee55))
+- **mui-paper:** add paper component ([4bf7776](https://github.com/Availity/element/commit/4bf7776a0f19e3566983e176901233a6292fa745))
+- **mui-paper:** refactor how props are spread ([ecddf69](https://github.com/Availity/element/commit/ecddf6908fdb69418705cb854aa84c419de3ee55))